+++ crates/objc2/CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `ffi::objc_object` is merged into `runtime::AnyObject`.
   - `ffi::IMP` is merged into `runtime::Imp`.
   - `ffi::objc_method` is merged into `runtime::Method`.
+  - `ffi::objc_ivar` is merged into `runtime::Ivar`.
 
 ### Fixed
 * Remove an incorrect assertion when adding protocols to classes in an unexpected